% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration-network-us-congress.R
\name{clean_string}
\alias{clean_string}
\title{Helper function to extract the names of legislators}
\usage{
clean_string(s)
}
\value{
String
}
\description{
Helper function to extract the names of legislators
}
