% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration-network-us-congress.R
\name{get_bio}
\alias{get_bio}
\title{get bios from http://bioguide.congress.gov/biosearch/biosearch1.asp}
\usage{
get_bio(pos)
}
\value{
bio df
}
\description{
get bios from http://bioguide.congress.gov/biosearch/biosearch1.asp
}
