% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hate-speech-detect.R
\name{hate_vocabulary}
\alias{hate_vocabulary}
\title{This function creates an English hate vocabulary based on hatebase.org}
\usage{
hate_vocabulary(k)
}
\arguments{
\item{API}{key}
}
\value{
A df of hate vocabulary
}
\description{
This function creates an English hate vocabulary based on hatebase.org
}
