% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collaboration-network-us-congress.R
\name{sponsors_list}
\alias{sponsors_list}
\title{Create a list of all sponsors containing bill name, sponsor and a vector of co-sponsors}
\usage{
sponsors_list(bill_dir)
}
\arguments{
\item{directory}{of bills}
}
\value{
sponsor list
}
\description{
Create a list of all sponsors containing bill name, sponsor and a vector of co-sponsors
}
